# Tugas-React/backend-tugas-reactjs/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "basePath": "/",
    "paths": {
...
        "/nilai": {
            "get": {
                "description": "Get All Nilai",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "nilai"
                ],
                "summary": "Get All Nilai",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Nilai"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new Nilai with the input payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "nilai"
                ],
                "summary": "Create a new Nilai",
                "parameters": [
                    {
                        "description": "Create Nilai",
                        "name": "nilai",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateNilaiInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Nilai"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/nilai/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a Nilai by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "nilai"
                ],
                "summary": "Get a Nilai by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Nilai ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Nilai"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a Nilai by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "nilai"
                ],
                "summary": "Delete a Nilai",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Nilai ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update a Nilai with the input payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "nilai"
                ],
                "summary": "Update a Nilai",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Nilai ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Nilai",
                        "name": "nilai",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateNilaiInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Nilai"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateNilaiInput": {
            "type": "object",
            "required": [
                "indeks",
                "mahasiswa_id",
                "mata_kuliah_id",
                "skor",
                "users_id"
            ],
            "properties": {
                "indeks": {
                    "type": "string"
                },
                "mahasiswa_id": {
                    "type": "integer"
                },
                "mata_kuliah_id": {
                    "type": "integer"
                },
                "skor": {
                    "type": "integer"
                },
                "users_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.Nilai": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "indeks": {
                    "type": "string"
                },
                "mahasiswaID": {
                    "type": "integer"
                },
                "mataKuliahID": {
                    "type": "integer"
                },
                "skor": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "usersID": {
                    "type": "integer"
                }
            }
        },
...
        "models.UpdateNilaiInput": {
            "type": "object",
            "properties": {
                "indeks": {
                    "type": "string"
                },
                "mahasiswa_id": {
                    "type": "integer"
                },
                "mata_kuliah_id": {
                    "type": "integer"
                },
                "skor": {
                    "type": "integer"
                },
                "users_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}